libexpr/parser: do not use isUri, check for scheme
Previously we would be using a hardcoded list of schemes that we
supported to check if a Path is actually a URI (that we can fetch).
This change decouples the parser from the actually implemented fetching.
Either something should be a valid path or a URI. If that URI is not
implemented then that is an eval failure. Just like before.